rishte-wali-aunty/pyproject.toml
[project]
name = "rishte-wali-aunty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "dotenv>=0.9.9",
    "openai-agents>=0.2.1",
]
